+++ src/index.js
@@ -1,7 +1,14 @@
 const express = require('express')
+const userRouter = require('./routes/users')
+
 const app = express()
-const port = 3000
+const port = 3001
+
 
 app.get('/', (req, res) => res.send('Hello World!'))
 
-app.listen(port, () => console.log(`Example app listening on port ${port}!`))
+//.app.use('/user', userRouter)
+
+const server = app.listen(port, () => console.log(`Example app listening on port ${port}!`))
+
+module.exports = server