+++ test/users.router.js
@@ -1,0 +1,16 @@
+const app = require('../src/index')
+
+describe('Users REST API', () => {
+
+  describe('POST /user', ()=> {
+
+    it('create a new user', (done)=> {
+
+    })
+  })
+
+  describe('GET /user', ()=> {
+    // TODO Create test for the get method
+
+  })
+})