+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+# Trunk
+
+* Create users REST API
+* Create unit test for users REST API ("create")
+
+# Version 1.1.0
+
+* Refactoring: move `src/users.js` to `src/controllers/users.js`
+* Create the method "create" of the module `users.js`
+* Create unit tests for the module `users.js`
+* Define the storage format
+
 ## Version 1.0.0
 
 * Create configuration module